+++ packages/terra-form-input/src/terra-dev-site/doc/example/IncompleteInputField.jsx
@@ -1,0 +1,18 @@
+import React from 'react';
+import InputField from 'terra-form-input/lib/InputField';
+
+const RequiredInputField = () => (
+  <InputField
+    inputId="requiredField"
+    label="Incomplete Input Field"
+    help="Note: This is help text"
+    type="text"
+    required
+    isIncomplete
+    inputAttrs={{
+      name: 'requiredField',
+    }}
+  />
+);
+
+export default RequiredInputField;