+++ packages/terra-form-select/src/terra-dev-site/doc/example/single/IncompleteSingleSelectField.jsx
@@ -1,0 +1,18 @@
+import React from 'react';
+import SingleSelectField from 'terra-form-select/lib/SingleSelectField';
+import classNames from 'classnames/bind';
+import styles from '../FormSelectDocCommon.module.scss';
+
+const cx = classNames.bind(styles);
+
+const SingleSelectFieldExample = () => (
+  <SingleSelectField required isIncomplete label="T-shirt size" placeholder="Select a size" selectId="tshirt-size-field" className={cx('form-select')}>
+    <SingleSelectField.Option value="xSmall" display="Extra Small" />
+    <SingleSelectField.Option value="small" display="Small" />
+    <SingleSelectField.Option value="medium" display="Medium" />
+    <SingleSelectField.Option value="large" display="Large" />
+    <SingleSelectField.Option value="xLarge" display="Extra Large" />
+  </SingleSelectField>
+);
+
+export default SingleSelectFieldExample;