+++ packages/terra-form-textarea/src/terra-dev-site/test/form-textarea/RowsTextarea.test.jsx
@@ -1,6 +1,14 @@
 import React from 'react';
+import classNames from 'classnames/bind';
 import Textarea from '../../../Textarea';
+import styles from './common/Textarea.test.module.scss';
 
-const textarea = () => <Textarea rows={15} size="small" ariaLabel="label" />;
+const cx = classNames.bind(styles);
+
+const textarea = () => (
+  <div className={cx('content-wrapper')}>
+    <Textarea rows={15} size="small" ariaLabel="label" />
+  </div>
+);
 
 export default textarea;