+++ packages/terra-form-textarea/CHANGELOG.md
@@ -6,6 +6,15 @@ Unreleased
 ### Changed
 * Updated Jest Snapshots of Textarea-Feild for the new props of Terra-Feild.
 
+3.29.0 - (October 14, 2019)
+------------------
+### Changed
+* Minor dependency version bump
+
+### Added
+* Added `placeholder` as a property to `<Textarea />` and `<TextareaField />`
+* Added `isIncomplete` prop to show incomplete stylings.
+
 3.28.0 - (October 3, 2019)
 ------------------
 ### Changed